fix clmul aarch64
